Change size of exnStack field to size_t. Since exnStack is added to stackBottom (a pointer), keeping it the same size as a pointer avoids the need for extension/truncation.
SVN r5558

+++ runtime/gc/gc_state.h
@@ -17,7 +17,7 @@ struct GC_state {
   pointer limit; /* limit = heap.start + heap.size */
   pointer stackTop; /* Top of stack in current thread. */
   pointer stackLimit; /* stackBottom + stackSize - maxFrameSize */
-  uint32_t exnStack;
+  size_t exnStack;
   /* Alphabetized fields follow. */
   size_t alignment; /* */
   bool amInGC;
